Allow Label.update_label to accept a dict
This allows both name and color to be changed.
Also factored out singledispatchmethod to a lib module as it's now used
in two modules.
